actor: Reorganize crate
